# api-spec/src/main/resources/openapi/sars-app.yaml
openapi: 3.0.0
info:
  title: PosAm Assignment
  description: API for managing animals, reservations, and users
  version: 1.0.0

paths:
  /animals:
    get:
      summary: Get all animals
      operationId: getAllAnimals
      responses:
        '200':
          description: List of all animals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
    post:
      summary: Create a new animal
      operationId: createAnimal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Animal'
      responses:
        '201':
          description: Animal created successfully
  /animals/{id}:
    get:
      summary: Get an animal by ID
      operationId: getAnimalById
      parameters:
        - name: id
          in: path    # Changed from 'query' to 'path'
          description: ID of the animal to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Animal found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '404':
          description: Animal not found
  /animals/available:
    get:
      summary: Get all animals within a time range
      operationId: getAllAnimalsWithinTimeRange
      parameters:
        - name: startTime
          in: query
          description: Start time of the time range (inclusive)
          required: true
          schema:
            type: string
            format: date-time-local
        - name: endTime
          in: query
          description: End time of the time range (exclusive)
          required: true
          schema:
            type: string
            format: date-time-local
      responses:
        '200':
          description: List of all animals within the time range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
  /reservations/{id}:
    get:
      summary: Get a reservation by ID
      operationId: getReservationById
      parameters:
        - name: id
          in: path
          description: ID of the reservation to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reservation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservation not found

  /reservations:
    get:
      summary: Get all reservations
      operationId: getAllReservations
      responses:
        '200':
          description: List of all reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
    post:
      summary: Create a new reservation
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
  /users:
    get:
      summary: Get a user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: query
          description: ID of the user to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  schemas:
    Animal:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Name of the animal
        race:
          $ref: '#/components/schemas/Race'
      required:
        - name
    Race:
      type: string
      enum:
        - LABRADOR
        - SHEPHERD
        - RETRIEVER
        - BULLDOG
        - BEAGLE
        - POODLE
        - ROTTWEILER
        - YORKSHIRE
        - SIAMESE
        - MAINE
        - RAGDOLL
        - BENGAL
        - SPHYNX
        - BRITISH
        - SCOTTISH
        - ABYSSINIAN
        - PERSIAN
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time-local
        startTime:
          type: string
          format: date-time-local
        endTime:
          type: string
          format: date-time-local
        status:
          $ref: '#/components/schemas/ReservationStatus'
        user:
          $ref: '#/components/schemas/User'
        animal:
          $ref: '#/components/schemas/Animal'
    ReservationStatus:
      type: string
      enum:
        - APPROVED
        - REJECTED
        - PENDING
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time-local
        name:
          type: string
        userName:
          type: string
        email:
          type: string
        credit:
          $ref: '#/components/schemas/Credit'
        role:
          $ref: '#/components/schemas/UserRole'
      required:
        - name
    UserRole:
      type: string
      enum:
        - ADMIN
        - DOGWALKER
    Credit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        count:
          type: integer